alpine

+++ modules/alpine_syscall_hooks/src/hook.c
@@ -363,7 +363,7 @@ int install_hooks(void) {
     syscall_table = lookup_syscall_table_by_file();// (long unsigned int*)kallsyms_lookup_name("sys_call_table");
     echo("sys_call_table address %lu\n", syscall_table);
     if(!syscall_table) return 0;
-    make_vm_rw(syscall_table);
+    make_vm_rw((long unsigned int)syscall_table);
 
     /** read
      */
@@ -385,7 +385,7 @@ int install_hooks(void) {
 void uninstall_hooks(void){
     if(!syscall_table) return ;
     echo("sys_call_table address %lu\n", syscall_table);
-    //make_vm_rw(syscall_table);
+    //make_vm_rw((long unsigned int)syscall_table);
     //syscall_table[__NR_read] = original_syscall_table[__NR_read];
     //syscall_table[__NR_getdents] = original_syscall_table[__NR_getdents];
     //syscall_table[__NR_getdents64] = original_syscall_table[__NR_getdents64];
